src/server/flask_server.py
"""
Flask Server for FiKing-Trader API
Handles requests for agent management, predictions, and data operations
"""

import os
import logging
from flask import Flask, request, jsonify
from flask_cors import CORS
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Initialize Flask app
app = Flask(__name__)
CORS(app)

# Configuration
app.config['JSON_SORT_KEYS'] = False
app.config['MAX_CONTENT_LENGTH'] = 16 * 1024 * 1024  # 16MB max request size


# ==================== Health Check ====================

@app.route('/health', methods=['GET'])
def health_check():
    """Health check endpoint to verify server is running"""
    return jsonify({
        'status': 'OK',
        'message': 'FiKing-Trader API is healthy'
    }), 200


@app.route('/status', methods=['GET'])
def get_status():
        """Get system status and statistics"""
        pass


# ==================== Model/Agent Management ====================

@app.route('/model/create', methods=['POST'])
def create_model():
    """
    Create and configure a forecasting model instance
    
    Request Body:
        - entity_name: Financial instrument identifier
        - time_frequency: Time resolution (1MIN, 5MIN, 1H, 1D, etc.)
        - observation_horizon: Number of past time steps for input
        - prediction_horizon: Number of time steps ahead to forecast
        - news_observation_horizon: (optional) Time steps for news data
        - news_retrieval_prompt: (optional) Natural language query for news
        - news_resources: (optional) List of news sources
        - confidence_level: (optional) Minimum confidence for signals
        - signal_frequency: Frequency for signal evaluation
    
    Response:
        - status: ACCEPTED or REJECTED
        - valid: Boolean indicating validation result
        - model_id: Unique identifier (if accepted)
        - message: Detailed explanation
        - input_summary: Human-readable configuration description*

    """
    body = request.get_json()
    service = body.get('service')
    
    pass


@app.route('/model/<model_id>/start', methods=['POST'])
def start_model(model_id):
    """
    Start/launch a trained model for continuous prediction
    
    Path Parameters:
        - model_id: Unique model identifier
    
    Response:
        - status: SUCCESS or FAILURE
        - message: Result description
    """
    pass


@app.route('/model/<model_id>/stop', methods=['POST'])
def stop_model(model_id):
    """
    Stop a running model
    
    Path Parameters:
        - model_id: Unique model identifier
    
    Response:
        - status: SUCCESS or FAILURE
        - message: Result description
    """
    pass


@app.route('/model/<model_id>/status', methods=['GET'])
def get_model_status(model_id):
    """
    Get status of a specific model
    
    Path Parameters:
        - model_id: Unique model identifier
    
    Response:
        - model_id: Model identifier
        - status: Current status (training, running, stopped, etc.)
        - is_trained: Boolean
        - is_running: Boolean
        - equity: Associated equity symbol
        - created_at: Creation timestamp
        - last_prediction_at: Last prediction timestamp
        - performance_metrics: Training/evaluation metrics
    """
    pass


@app.route('/model/<model_id>', methods=['DELETE'])
def delete_model(model_id):
    """
    Delete a model and its associated data
    
    Path Parameters:
        - model_id: Unique model identifier
    
    Response:
        - status: SUCCESS or FAILURE
        - message: Result description
    """
    pass


@app.route('/models', methods=['GET'])
def list_models():
    """
    List all models
    
    Query Parameters:
        - equity: (optional) Filter by equity symbol
        - status: (optional) Filter by status
        - limit: (optional) Maximum number of results
        - offset: (optional) Pagination offset
    
    Response:
        - models: List of model objects
        - total: Total number of models
    """
    pass


# ==================== Inference/Prediction ====================

@app.route('/model/<model_id>/predict', methods=['POST'])
def predict(model_id):
    """
    Request on-demand prediction from a model
    
    Path Parameters:
        - model_id: Unique model identifier
    
    Request Body:
        - input_data: Features for prediction
        - confidence_level: (optional) Confidence threshold
    
    Response:
        - prediction: Model prediction result
        - confidence: Prediction confidence
        - timestamp: Prediction timestamp
    """
    pass


@app.route('/model/<model_id>/signals', methods=['GET'])
def get_signals(model_id):
    """
    Get recent trading signals from a model
    
    Path Parameters:
        - model_id: Unique model identifier
    
    Query Parameters:
        - limit: (optional) Maximum number of signals
        - start_date: (optional) Filter from date
        - end_date: (optional) Filter to date
    
    Response:
        - signals: List of signal objects
        - total: Total number of signals
    """
    pass


# ==================== Data Management ====================

@app.route('/data/timeseries', methods=['POST'])
def upload_timeseries_data():
    """
    Upload time series data for an equity
    
    Request Body:
        - equity: Equity symbol
        - frequency: Time frequency
        - data: Array of OHLCV data points
    
    Response:
        - status: SUCCESS or FAILURE
        - records_inserted: Number of records inserted
        - message: Result description
    """
    pass


@app.route('/data/timeseries/<equity>', methods=['GET'])
def get_timeseries_data(equity):
    """
    Retrieve time series data for an equity
    
    Path Parameters:
        - equity: Equity symbol
    
    Query Parameters:
        - frequency: Time frequency (required)
        - start_date: (optional) Filter from date
        - end_date: (optional) Filter to date
        - limit: (optional) Maximum number of records
    
    Response:
        - equity: Equity symbol
        - frequency: Time frequency
        - data: Array of OHLCV data points
        - total: Total number of records
    """
    pass


@app.route('/data/equities', methods=['GET'])
def list_equities():
    """
    List all equities with available data
    
    Query Parameters:
        - frequency: (optional) Filter by frequency
    
    Response:
        - equities: List of equity symbols
        - total: Total number of equities
    """
    pass


@app.route('/data/live/<equity>', methods=['GET'])
def get_live_price(equity):
    """
    Get live price data for an equity
    
    Path Parameters:
        - equity: Equity symbol
    
    Response:
        - equity: Equity symbol
        - price: Current price data (OHLCV)
        - timestamp: Data timestamp
    """
    pass


# ==================== Training ====================

@app.route('/model/<model_id>/train', methods=['POST'])
def train_model(model_id):
    """
    Trigger model training
    
    Path Parameters:
        - model_id: Unique model identifier
    
    Request Body:
        - training_params: (optional) Training parameters override
    
    Response:
        - status: SUCCESS or FAILURE
        - message: Result description
        - job_id: Training job identifier (if async)
    """
    pass


@app.route('/model/<model_id>/training/status', methods=['GET'])
def get_training_status(model_id):
    """
    Get training status for a model
    
    Path Parameters:
        - model_id: Unique model identifier
    
    Response:
        - status: Training status (pending, running, completed, failed)
        - progress: Training progress percentage
        - current_epoch: Current training epoch
        - total_epochs: Total number of epochs
        - metrics: Current training metrics
    """
    pass


# ==================== Configuration ====================

@app.route('/config/entities', methods=['GET'])
def get_entities():
    """
    Get list of supported entities from config
    
    Response:
        - entities: List of entity configurations
    """
    pass


@app.route('/config/frequencies', methods=['GET'])
def get_time_frequencies():
    """
    Get list of supported time frequencies
    
    Response:
        - frequencies: List of time frequency configurations
    """
    pass


# ==================== Error Handlers ====================

@app.errorhandler(400)
def bad_request(error):
    """Handle bad request errors"""
    return jsonify({
        'status': 'ERROR',
        'message': 'Bad request',
        'error': str(error)
    }), 400


@app.errorhandler(404)
def not_found(error):
    """Handle not found errors"""
    return jsonify({
        'status': 'ERROR',
        'message': 'Resource not found',
        'error': str(error)
    }), 404


@app.errorhandler(500)
def internal_error(error):
    """Handle internal server errors"""
    logger.error(f"Internal server error: {str(error)}")
    return jsonify({
        'status': 'ERROR',
        'message': 'Internal server error',
        'error': str(error)
    }), 500


# ==================== Main ====================

def run_server(host='0.0.0.0', port=5000, debug=False):
    """
    Run the Flask server
    
    Args:
        host: Host address
        port: Port number
        debug: Debug mode flag
    """
    logger.info(f"Starting Flask server on {host}:{port}")
    app.run(host=host, port=port, debug=debug)


if __name__ == '__main__':
    # Get configuration from environment
    host = os.getenv('FLASK_HOST', '0.0.0.0')
    port = int(os.getenv('FLASK_PORT', 5000))
    debug = os.getenv('FLASK_DEBUG', 'False').lower() == 'true'
    
    run_server(host=host, port=port, debug=debug)
