src/server/requirements-flask.txt
# Requirements for Flask Server
# Install with: pip install -r requirements-flask.txt

Flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=0.19.0
